feat(major_overhaul): tonnes up updates to improve firmware including:
* added calculations to get gyroscope and accel values (including without gravity), OSC output for these values, and a function to convert these values (int16_t) to [-1,1]

* found that gyroscope was set to sensativity of +- 2000. therefore, divided gyro value by 2000. now is [-1, 1]

* switched the x and z axis output for the OSC gyro message to match the ordering of YPR

* adjusted max object to better match the ordering of YPR, Gyro, Accel

* got brightness implemented. doesnt work very well because of the way the Neopixel library was written. Basically the pixel values in ram are scaled and then pushed out so they cant come back if set to zero

* added HSL control via OSC endpoint

* basic motion sensing works. needs hysterisis

* added data container for IMU

* fixed fmod issue with shitty static function

* adjusted all the osc outputs to get values from new motionstate object

* commented out YPR stability. will fix another time

* updated maxpatch with API endpoints

* made some max examples for multiple cubes

* did some minor code cleanup, removing functions from main that have been implemented into smaller works.

added some more max application example

got code working on new build system. updated the readme to account for proper library setups.

fixed the issue with choosing side. turns out int8_t was being cast incorrectly to int or something. changed type of _side in MotionState and the return type of whichSide() in the same class to fix this issue.

began a new class to deal with persistant rotation state per side of cube

removed math library from MotionState (technically in last commit) and began to add SideValue to MotionState.

persistant sides now work!

fixed the max external so it no longer requires an ID

edited the python osc parse script to output to screen in a format whcih will work for csv. the data is so small i am doing it by hand

edited python record to be easier to read

added a new example patch for max which makes the dreamIO control parameters of a simple synth

fixed bug that would cause OSC transmission to happen constantly until the IMU interrupt.

changed the API so that setting the update interval to 0 makes the transmission stop

adjusted the rotation threshold. this should make the side values more stable

update readme for formatting